ontouchup delay before transition from walk to idle @WXR-012
